Fixed #33826 -- Fixed RedisCache.set_many()/delete_many() crash with an empty list.

+++ django/core/cache/backends/redis.py
@@ -214,6 +214,8 @@ class RedisCache(BaseCache):
         return self._cache.incr(key, delta)
 
     def set_many(self, data, timeout=DEFAULT_TIMEOUT, version=None):
+        if not data:
+            return []
         safe_data = {}
         for key, value in data.items():
             key = self.make_and_validate_key(key, version=version)
@@ -222,6 +224,8 @@ class RedisCache(BaseCache):
         return []
 
     def delete_many(self, keys, version=None):
+        if not keys:
+            return
         safe_keys = [self.make_and_validate_key(key, version=version) for key in keys]
         self._cache.delete_many(safe_keys)
